refactor analytics
