+++ src/shared/analytics/analytics.tsx
@@ -1,8 +1,8 @@
 import React, { useEffect, ComponentType } from 'react';
-import ReactGA, { FieldsObject, EventArgs } from 'react-ga';
+import ReactGA, { FieldsObject } from 'react-ga';
 import { RouteComponentProps } from 'react-router-dom';
 
-import { idleCallback } from './shared/utils/idleCallback';
+import { idleCallback } from '../utils/idleCallback';
 
 const config = [
   {
@@ -35,20 +35,4 @@ export const withRouteTracker = <P extends RouteComponentProps>(
 
     return <WrappedComponent {...props} />;
   };
-};
-
-export const analyticsEvent = (options: EventArgs) => {
-  const event = () => {
-    if (process.env.NODE_ENV === 'production') {
-      ReactGA.event({ ...options });
-    }
-  };
-  idleCallback(() => event());
-};
-
-export const eventActions = {
-  clicked: (text: string) => `Clicked ${text}`
-};
-export const eventCategories = {
-  user: 'User'
 };
